+++ package/sdk/src/lib/config.ts
@@ -52,7 +52,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "1.0.0",
-  sdkVersion: "0.1.6",
-  genVersion: "2.438.2",
-  userAgent: "speakeasy-sdk/typescript 0.1.6 2.438.2 1.0.0 ryan-total-test-act",
+  sdkVersion: "0.1.7",
+  genVersion: "2.438.3",
+  userAgent: "speakeasy-sdk/typescript 0.1.7 2.438.3 1.0.0 ryan-total-test-act",
 } as const;